+++ src/components/CarbonItems.tsx
@@ -1,4 +1,5 @@
-import { Box, Center, GridItem } from '@chakra-ui/layout';
+import { Box, Center, Grid, GridItem } from '@chakra-ui/layout';
+import { svgBackgroundImage } from '@guardian/src-helpers';
 import type { SectionType } from '../sections/Sections';
 import { sections } from '../sections/Sections';
 
@@ -8,6 +9,20 @@ export type CarbonItem = {
 	emoji: string;
 	section: SectionType;
 };
+
+const gridSize = 24;
+const gapSize = 2;
+const backgroundSize = gridSize + gapSize;
+
+const grid = svgBackgroundImage(`
+		<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 ${backgroundSize} ${backgroundSize}">
+			<g fill="#abc">
+				<circle cx="0" cy="0" r="1" />
+				<circle cx="${backgroundSize}" cy="0" r="1" />
+				<circle cx="0" cy="${backgroundSize}" r="1" />
+				<circle cx="${backgroundSize}" cy="${backgroundSize}" r="1" />
+			</g>
+		</svg>`);
 
 const approximate = (min: number, max: number): number =>
 	Math.floor(Math.random() * (max - min)) + min;
@@ -38,14 +53,30 @@ const CarbonItemBox = ({ item }: { item: CarbonItem }): JSX.Element => {
 			<Center h={'100%'} flexDirection="column">
 				<Box
 					fontSize={`${Math.min(6, Math.min(rows, cols))}rem`}
-					lineHeight="0.75"
+					lineHeight="1"
 				>
 					{item.emoji}
 				</Box>
-				<Box>{item.name}</Box>
+				<Box textAlign="center">{item.name}</Box>
 			</Center>
 		</GridItem>
 	);
 };
 
-export { CarbonItemBox, boxFromCarbon, approximate };
+const CarbonItemWithGrid = ({ item }: { item: CarbonItem }) => {
+	return (
+		<Grid>
+			<CarbonItemBox item={item} />;
+		</Grid>
+	);
+};
+
+export {
+	CarbonItemBox,
+	boxFromCarbon,
+	approximate,
+	gridSize,
+	gapSize,
+	backgroundSize,
+	grid,
+};
